refactor: remove duplicated logic for checking bin in path
